Move command substitution parser to separate module
